@@ -17,14 +17,19 @@
         "doctrine/common": "^2.7.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.4.6",
+        "doctrine/coding-standard": "^4.0",
+        "phpunit/phpunit": "^7.0",
         "phpunit/phpunit-mock-objects": "!=3.2.4,!=3.2.5",
-        "symfony/console": "2.*||^3.0"
+        "symfony/console": "^2.0.5||^3.0",
+        "symfony/phpunit-bridge": "^3.4.5|^4.0.5"
     },
     "suggest": {
         "symfony/console": "For helpful console commands such as SQL execution and import of files."
     },
     "bin": ["bin/doctrine-dbal"],
+    "config": {
+        "sort-packages": true
+    },
     "autoload": {
         "psr-0": { "Doctrine\\DBAL\\": "lib/" }
     },
@@ -33,10 +38,10 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.6.x-dev"
+            "dev-master": "2.7.x-dev"
         }
     },
     "archive": {
-        "exclude": ["!vendor", "tests", "*phpunit.xml", ".travis.yml", "build.xml", "build.properties", "composer.phar"]
+        "exclude": ["!vendor", "tests", "*phpunit.xml", ".appveyor.yml", ".travis.yml", "build.xml", "build.properties", "composer.phar"]
     }
 }
