@@ -33,7 +33,7 @@ return [
       | This tells about aplication current version.
       |
      */
-    'version' => 'Community 1.9.7-BETA',
+    'version' => 'Community 1.10',
     /*
       |--------------------------------------------------------------------------
       | Application Name
