@@ -7,7 +7,6 @@ $extensions = [
     'imap',
     'mbstring',
     'mcrypt',
-    'mysql',
     'openssl',
     'tokenizer',
     'zip',
@@ -15,7 +14,7 @@ $extensions = [
     'mysqli',
     'bcmath',
     'iconv',
-    'XML-DOM', //for HTML email processing
+    'XML', //for HTML email processing
     'json',
     //'ioncube_loader_dar_5.6',
 ];
@@ -65,10 +64,10 @@ $extensions = [
                 echo '</tr>';
             }
             echo '<tr>';
-            if (phpversion() >= 5.6) {
+            if (phpversion() >= 7.1) {
                 echo "<td>PHP Version</td>  <td style='color:green'>".phpversion().'</td>';
             } else {
-                echo "<td>PHP Version</td>  <td style='color:red'>".phpversion().'<p>Please upgrade PHP Version to 5.6+ </p></td>';
+                echo "<td>PHP Version</td>  <td style='color:red'>".phpversion().'<p>Please upgrade PHP Version to 7.1+ </p></td>';
             }
             echo '</tr>';
             echo '<tr>';
