@@ -13,6 +13,6 @@ return [
       |
      */
 
-    'previous' => '&laquo; zurück',
-    'next'     => 'weiter &raquo;',
+    'previous' => '&laquo; Zurück',
+    'next'     => 'Weiter &raquo;',
 ];
