@@ -13,66 +13,67 @@ return [
       |
      */
 
-    'accepted'   => 'Das :attribute muss akzeptiert werden.',
+    'accepted'   => 'Das :attribute muss akzeptiert sein.',
     'active_url' => 'Das :attribute ist keine gültige URL.',
-    'after'      => 'Das :attribute muss ein Datum sein, nachdem :date.',
-    'alpha'      => 'Das :attribute dürfen nur Buchstaben sein.',
-    'alpha_dash' => 'Das :attribute darf nur Buchstaben, Zahlen und Bindestriche enthalten.',
-    'alpha_num'  => 'Das :attribute darf nur Buchstaben und Zahlen enthalten.',
-    'array'      => 'Das :attribute muss ein Array sein.',
-    'before'     => 'Das :attribute muss ein Datum sein, bevor :date.',
+    'after'      => 'Das :attribute muss ein Datum nach :date sein.',
+    'alpha'      => 'Das :attribute darf nur Buchstaben enthalten.',
+    'alpha_dash' => 'Das :attribute darf nur Buchstaben, Ziffern und Minus enthalten.',
+    'alpha_num'  => 'Das :attribute darf nur Ziffern und Buchstaben enthalten.',
+    'array'      => 'Das :attribute muss ein Arrey sein.',
+    'before'     => 'Das :attribute muss ein Datum vor :date sein.',
     'between'    => [
-        'numeric' => 'Das :attribute muss zwischen :min und :max.',
-        'file'    => 'Das :attribute muss zwischen :min und :max Kilobyte.',
-        'string'  => 'Das :attribute muss zwischen :min und :max Zeichen.',
-        'array'   => 'Das :attribute muss genau zwischen :min und :max Anzahl.',
+        'numeric' => 'Das :attribute muss zwischen :min und :max liegen.',
+        'file'    => 'Das :attribute muss zwischen :min und :max kilobytes sein.',
+        'string'  => 'Das :attribute muss zwischen :min und :max Buchstaben sein.',
+        'array'   => 'Das :attribute muss zwischen :min und :max Elemente haben.',
     ],
-    'boolean'        => 'Das :attribute Feld muss wahr oder falsch sein.',
+    'boolean'        => 'Das :attribute muss wahr oder falsch sein.',
     'confirmed'      => 'Das :attribute Bestätigung stimmt nicht überein.',
     'date'           => 'Das :attribute ist kein gültiges Datum.',
-    'date_format'    => 'Das :attribute passt nicht mit dem Format überein :format.',
-    'different'      => 'Das :attribute und :other müssen unterschiedlich sein.',
-    'digits'         => 'Das :attribute muss sein :digits Ziffern.',
-    'digits_between' => 'Das :attribute muss zwischen :min und :max Ziffern.',
+    'date_format'    => 'Das :attribute verstößt gegen das Format :format.',
+    'different'      => 'Das :attribute und :other müssen sich unterscheiden.',
+    'digits'         => 'Das :attribute müssen :digits Ziffern.',
+    'digits_between' => 'Das :attribute muss zwischen :min und :max Ziffern haben.',
     'email'          => 'Das :attribute muss eine gültige E-Mail-Adresse sein.',
-    'filled'         => 'Das :attribute ist erforderlich.',
-    'exists'         => 'Das ausgewählte :attribute ist ungültig.',
-    'image'          => 'Das :attribute muß ein Bild sein.',
-    'in'             => 'Das ausgewählte :attribute ist ungültig.',
-    'integer'        => 'Das :attribute muss eine ganze Zahl sein.',
-    'ip'             => 'Das :attribute muss eine gültige IP-Adresse sein.',
+    'filled'         => 'Das :attribute Feld ist Pflicht.',
+    'exists'         => 'Das gewählte :attribute ungültg.',
+    'image'          => 'Das :attribute muss ein Bild sein.',
+    'in'             => 'Das gewählte :attribute ist ungültig.',
+    'integer'        => 'Das :attribute muss ein Integer sein.',
+    'ip'             => 'Das :attribute muss eine gültige IP-Adress sein.',
     'max'            => [
         'numeric' => 'Das :attribute darf nicht größer sein als :max.',
-        'file'    => 'Das :attribute darf nicht größer sein als :max Kilobytes.',
+        'file'    => 'Das :attribute darf nicht größer sein als :max kilobytes.',
         'string'  => 'Das :attribute darf nicht größer sein als :max Zeichen.',
-        'array'   => 'Das :attribute kann nicht mehr sein als :max Anzahl.',
+        'array'   => 'Das :attribute darf nicht mehr als :max Elemente haben.',
     ],
-    'mimes' => 'Das :attribute must be a file of type: :values.',
+    'mimes' => 'Das :attribute muss eine Datei vom Typ :values sei.',
     'min'   => [
-        'numeric' => 'Das :attribute muss mindestens :min.',
-        'file'    => 'Das :attribute muss mindestens :min Kilobytes.',
-        'string'  => 'Das :attribute muss mindestens :min Zeichen.',
-        'array'   => 'Das :attribute darf nicht kleiner sein als :min Anzahl.',
+        'numeric' => 'Das :attribute muss mindesten :min sein.',
+        'file'    => 'Das :attribute muss mind. :min kilobytes groß sein.',
+        'string'  => 'Das :attribute muss mind. :min Zeichen groß sein.',
+        'array'   => 'Das :attribute muss mindesten :min Elemente haben.',
+
     ],
     'not_in'               => 'Das ausgewählte :attribute ist ungültig.',
-    'numeric'              => 'Das :attribute muss eine Nummer sein.',
+    'numeric'              => 'Das :attribute muss eine Zahl sein.',
     'regex'                => 'Das :attribute Format ist ungültig.',
-    'required'             => 'Das :attribute ist erforderlich.',
-    'required_if'          => 'Das :attribute Feld ist erforderlich, wenn :other ist :value.',
-    'required_with'        => 'Das :attribute Feld ist erforderlich, wenn :values ist anwesend.',
-    'required_with_all'    => 'Das :attribute Feld ist erforderlich, wenn :values ist anwesend.',
-    'required_without'     => 'Das :attribute Feld ist erforderlich, wenn :values ist nicht vorhanden.',
-    'required_without_all' => 'Das :attribute ist erforderlich, wenn keine der :values sind anwesend.',
-    'same'                 => 'Das :attribute und :other muss passen.',
+    'required'             => 'Das :attribute Feld wird benötigt.',
+    'required_if'          => 'Das :attribute Feld wird benötigt wenn :other :value. ist.',
+    'required_with'        => 'Das :attribute Feld wird benötigt wenn :values vorhanden ist.',
+    'required_with_all'    => 'Das :attribute Feld wird benötigt wenn :values vorhanden ist.',
+    'required_without'     => 'Das :attribute Feld wird benötigt wenn :values nicht vorhanden ist.',
+    'required_without_all' => 'Das :attribute Feld wird benötigt wenn keines von :values verfügbar ist.',
+    'same'                 => 'Das :attribute und :other müssen übereinstimmen.',
     'size'                 => [
-        'numeric' => 'Das :attribute muss sein :size.',
-        'file'    => 'Das :attribute muss sein :size Kilobytes.',
-        'string'  => 'Das :attribute muss sein :size Zeichen.',
-        'array'   => 'Das :attribute muss enthalten :size Anzahl.',
+        'numeric' => 'Das :attribute muss :size groß sein.',
+        'file'    => 'Das :attribute muss :size Kilobytes groß sein.',
+        'string'  => 'Das :attribute muss :size Zeichen lang sein.',
+        'array'   => 'Das :attribute muss :size Elemente beinhalten.',
     ],
-    'unique'   => 'Das :attribute bereits aufgenommen wurde.',
+    'unique'   => 'Das :attribute wurde bereits verwendet.',
     'url'      => 'Das :attribute Format ist ungültig.',
-    'timezone' => 'Das :attribute muss eine gültige Zone sein.',
+    'timezone' => 'Das :attribute muss eine gültige Zeitzone sein.',
     /*
       |--------------------------------------------------------------------------
       | Custom Validation Language Lines
